--- code_rasp-TFlite_bookwarm/braulio/draft_3.py
@@ -1,6 +1,0 @@
-global nivel
-
-nivel = "bajo"
-
-while True:
-    nivel = str(input("inserte nivel deseado:"))
