+++ code_rasp-TFlite_bookwarm/braulio/test_semaforos/draft_3.py
@@ -1,0 +1,6 @@
+global nivel
+
+nivel = "bajo"
+
+while True:
+    nivel = str(input("inserte nivel deseado:"))
